Add conditional copying definitions to awi-ui
Currently, the awi-grpc repository doesn't provide any npm package
which could be used to set a dependency on the protos from awi-ui
repository.

Previously, the generate option from Makefile would copy generated
files to the awi-ui repository which was assumed to be present in
the same directory as awi-grpc repository. This change checks if
the awi-ui repository exists first.