Add region field to Network Domain proto
This change adds a region field for Network Domains as multiple
resources look for VPCs using Network Domain Listing and then
they need to check what is the region associated with them.

Apart from the performance (sending additional requests to check
the already known region) there is a risk of using invalid
Network Domain if there are multiple VPCs with the same ID in
different regions.